Updated Model Item category user

+++ controllers/categoryController.js
@@ -25,15 +25,63 @@ module.exports.getCategoryProducts = async (req, res) => {
   }
   
 };
-
-module.exports.deleteCategory= async (req, res) => {
+module.exports.addSubcategory = async (req, res) => {
   const { categoryId } = req.params;
+  const { subcategoryName } = req.body;
+
   try {
-    await Category.findByIdAndDelete(categoryId);
+    const category = await Category.findById(categoryId);
+
+    if (!category) {
+      return res.status(404).json({ error: 'Category not found' });
+    }
+
+    // Check if the subcategory already exists in the category's subcategories array
+    if (category.subcategories.includes(subcategoryName)) {
+      return res.status(400).json({ error: 'Subcategory already exists' });
+    }
+
+    // Add the new subcategory to the category's subcategories array
+    category.subcategories.push(subcategoryName);
+    await category.save();
+
+    res.status(201).json({ message: 'Subcategory added successfully' });
+  } catch (error) {
+    console.error('Error adding subcategory:', error);
+    res.status(500).json({ error: 'An error occurred while adding the subcategory' });
+  }
+};
+
+module.exports.deleteCategory = async (req, res) => {
+  const { categoryId } = req.params;
+  const { subcategoryName } = req.query;
+
+  try {
+    const category = await Category.findById(categoryId);
+
+    if (!category) {
+      return res.status(404).json({ error: 'Category not found' });
+    }
+
+    if (subcategoryName) {
+      // Delete a subcategory
+      const updatedSubcategories = category.subcategories.filter(
+        (subcategory) => subcategory !== subcategoryName
+      );
+
+      category.subcategories = updatedSubcategories;
+    } else {
+      // Delete the entire category
+      await Category.findByIdAndDelete(categoryId);
+    }
+
+    await category.save();
+
     res.sendStatus(204);
   } catch (error) {
     console.error('Error deleting category:', error);
     res.status(500).json({ error: 'An error occurred while deleting the category' });
   }
-}
+};
+
